Release 4.8.0
- [API] Deprecate all gQUIC versions.
- [FIX] http1x: check size of Cookie: header.
- [FIX] Add checks to gQUIC client session resume code.
- [FIX] 32-bit parser bounds checks.
- [FIX] Use ls-qpack v2.6.3 to pick up a bunch of fixes.
- [DOCS] FAQ: example programs do work with google.com etc.

+++ src/liblsquic/lsquic_parse_gquic_common.c
@@ -30,7 +30,7 @@
 #include "lsquic_logger.h"
 
 #define CHECK_SPACE(need, pstart, pend)  \
-    do { if ((intptr_t) (need) > ((pend) - (pstart))) { return -1; } } while (0)
+    do { if ((uint64_t) (need) > (uint64_t) ((pend) - (pstart))) { return -1; } } while (0)
 
 /* This partially parses `packet_in' and returns 0 if in case it succeeded and
  * -1 on failure.
